Parse Meat as a long in mall purchase request. Allow purchasing way too many raffle tickets. (#788)

+++ src/net/sourceforge/kolmafia/request/RaffleRequest.java
@@ -96,7 +96,7 @@ public class RaffleRequest extends GenericRequest {
     }
 
     int quantity = StringUtilities.parseInt(matcher.group(1));
-    int cost = 10000 * quantity;
+    long cost = 10000 * quantity;
 
     if (where.equals(RaffleSource.STORAGE.toString())) {
       KoLCharacter.setStorageMeat(KoLCharacter.getStorageMeat() - cost);
